node v12: deprecate ts3.3
Definitely Typed doesn't support TS 3.3 any longer.
