+++ types/node/v12/package.json
@@ -2,11 +2,6 @@
     "private": true,
     "types": "index",
     "typesVersions": {
-        "<=3.3": {
-            "*": [
-                "ts3.3/*"
-            ]
-        },
         "<=3.6": {
             "*": [
                 "ts3.6/*"
