--- types/node/v12/ts3.3/globals.global.d.ts
@@ -1,1 +1,0 @@
-declare var global: NodeJS.Global;
