--- types/node/v12/ts3.3/tslint.json
@@ -1,8 +1,0 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "ban-types": false,
-        "no-bad-reference": false,
-        "no-single-declare-module": false
-    }
-}
