+++ types/node/v12/ts3.6/index.d.ts
@@ -3,5 +3,4 @@
 // Typically type modifications should be made in base.d.ts instead of here
 
 /// <reference path="base.d.ts" />
-
-/// <reference path="../ts3.3/assert.d.ts" />
+/// <reference path="assert.d.ts" />
