+++ types/node/v12/ts3.6/tslint.json
@@ -1,6 +1,8 @@
 {
-  "extends": "dtslint/dt.json",
-  "rules": {
-    "no-bad-reference": false
-  }
+    "extends": "dtslint/dt.json",
+    "rules": {
+        "ban-types": false,
+        "no-bad-reference": false,
+        "no-single-declare-module": false
+    }
 }
